[ext-225] add zod validation for line item for existing product
